Exception added

+++ src/main/java/ru/vorobev/spring_boot_library/servises/BooksServices.java
@@ -1,17 +1,15 @@
 package ru.vorobev.spring_boot_library.servises;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.vorobev.spring_boot_library.exeptions.NotFoundException;
-import ru.vorobev.spring_boot_library.exeptions.ResourceNotFoundExceptions;
 import ru.vorobev.spring_boot_library.models.Book;
 import ru.vorobev.spring_boot_library.repositories.BookRepository;
 
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.List;
 
+@Slf4j
 @Service
 public class BooksServices {
     private BookRepository bookRepository;
@@ -26,10 +24,6 @@ public class BooksServices {
     }
 
     public List<Book> findAllBooks() {
-        List<Book> bookList = bookRepository.findAll();
-        for (Book book : bookList) {
-            System.out.println(book);
-        }
         return bookRepository.findAll();
     }
 
@@ -37,27 +31,37 @@ public class BooksServices {
         System.out.println(bookRepository.findById(id));
         System.out.println("findById");
         return bookRepository.findById(id).orElseThrow(() ->
-                new NotFoundException("Книга не найдена "+new Timestamp(System.currentTimeMillis())));
+                new NotFoundException("Book "+id +" not found "));
     }
 
     @Transactional
     public Book updateBook(Book book, Integer id) {
+        log.info("updateBook " + book.toString());
         Book updatedBook = bookRepository.findById(id).orElseThrow(() ->
-                new NotFoundException("Книга не найдена"));
+                new NotFoundException("Book "+id +" not found "));
         updatedBook.setAuthor(book.getAuthor());
+        updatedBook.setTimeBookTaking(book.getTimeBookTaking());
         updatedBook.setBookIsTaken(book.getBookIsTaken());
         updatedBook.setTitle(book.getTitle());
         updatedBook.setYearOfPublication(book.getYearOfPublication());
         updatedBook.setPerson(book.getPerson());
-        bookRepository.save(updatedBook);
-        return updatedBook;
+        return bookRepository.save(updatedBook);
+
     }
 
-    public List<Book> bookIsTaken() {
+    //TODO add in controller Books
+    public List<Book> findBookIsTaken() {
         return bookRepository.findBooksByBookIsTakenTrue();
     }
 
     public void deleteBookById(Integer id) {
+        if (bookRepository.findById(id).isEmpty())
+            throw new NotFoundException("Book "+id +" not found ");
         bookRepository.deleteById(id);
+    }
+
+    public Book bookIsTaken(Book book, Integer id) {
+
+        return null;
     }
 }
